renamed sciety to enhanced_preprints package
