+++ data_hub_api/enhanced_preprints/docmaps/sql/docmaps_index.sql
